requirements.txt
Flask
faiss-cpu
sentence-transformers
google-generativeai
PyPDF2
requests